lambda/marketing_cost_facebook/requirements.txt
aiohttp==3.11.13
aioboto3==11.1.1
facebook-business==21.0.0